pyproject.toml
[project]
name = "saliency-alignment"
version = "0.1.0"
description = "Fine-tuning video-language models with saliency alignment term for improved grounding."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.4.1",
    "hydra-core>=1.3.2",
    "lightning>=2.5.6",
    "pillow>=12.0.0",
    "pycocotools>=2.0.10",
    "transformers>=4.57.1",
    "wandb>=0.23.0",
    "deepspeed>=0.17",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.4",
]

[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["finetune"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]  # basics: style, flakes, isort, pyupgrade, bugbear
ignore = ["E501"]                    # let the formatter handle wrapping

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.mypy]
python_version = "3.11"
files = ["finetune"]
exclude = ["\\.venv", "env", "build", "dist"]
follow_imports = "skip"  # don't analyze third-party libs
ignore_missing_imports = true
